Fix Data Types, Errors - Added Documentation
Data Types
Updated data types for several integer values from strings to integers.
Updated MySQL columns to match updated data types. This change was made
to allow for the appropriate operations when sorting and filtering.

Dictionary Keys
Updated all dict keys to match column names from MySQL tables. This
was done to provide an easier transition in the future to dynamic
queries. Right now, the query string has to list columns to insert in
the same order as the keys were created on the dict. With matching
keys to table columns, I can update the query to use keys as columns
to insert, which will match the order of values to those keys.

Parameter Error When Uploading
When uploading data from various reports, I was getting an error
stating there weren't enough parameters. This was due to sections
of my code only adding a dict key/value if the report included data.
This meant the value would be skipped if the field was blank on the
report which led to fewer parameters than necessary. Now the
process will always insert a None value if the report string is empty.
After making this change, all uploads run without issue.

Empty Strings
I was inserting empty strings when a value from the respective report
was an empty string. This caused issues with converting strings to ints
or dates when the report had an empty string. Now a None value is set
wherever values are omitted in each report.

New Case in Primary Owner and Premises Addr. Field
I ran into a new case in one of the reports where there was no account
name or premises address, but there was an owner. The report still lists
the comma separator between street and city even if the values are blank
which left my array with an element contatining only a comma. Logic
was added to check if the address was a single comma, and if so,
it would set all address fields to None.

File Error Handling
When opening the CSV report, I have now added error handling in the case
that the file does not exist. The process will not crash if the file
does not exist. This can happen regularly as there are not always
reports available on the ABC website (such as no new applications on a
given day).

Comments and Doc Strings
Docstrings have been added to all functions explaining takes and returns
as well as general overview of the functions pupose. Comments were
added to segments of code that would benefit from additional context,
such as justifying certain decisions that were made.

Testing
Last commit had not tested new_applications process as the report
was not available through ABC. I have a copy of all three reports now
and tested/confirmed the functionality of this function.