Implemented Dynamic Resize to DetailsList - Now save window height to state on render/resize - Use height to adjust size of DL content wrapper
Fluent UI DetailsList has a wrapper called ms-Viewport that prevents the
content wrapper from filling 100% of the parent container. ms-Viewport
is immutable, so my options were limited to either a fixed size, or
calculating the height of the window. There are a couple of things to
note with this solution:
- vh modifier needs to be adjusted based on size of media
- Root webpart styles need to include size larger than DL for continuity