Moved Pagination/Filtering/Handling to Front End - APIs no longer serve paginated data - API calls made once for today/all per report render - Pagination handled in DataList component - Filtering, sorting, and today/all toggles updated to use pagination - Apply/Reset filters button/functionality implemented
Tested all cases and combinations of filters/sorts and only found one
issues that doesn't necessarily affect functionality. With sorted
data, if a filter is applied, the sort has to be reapplied. Then when
the filters are reset, the column sort icons still appear, but data is
not sorted correctly. This needs to be updated for consistency.